# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Pharmacy",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of pharmacy app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://pharmacy-backend-kwps.onrender.com"
    }
  ],
...
  "paths": {
...
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Get user cart",
        "operationId": "getCart",
        "description": "Get user cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found cart!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/get"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add products to cart",
        "operationId": "addProducts",
        "description": "Add products to cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product was successfully added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Product was successfully added"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/get"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "get": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "items",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "description": "Cart id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "userId": {
            "description": "User id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "items": {
            "description": "Products chosen by user",
            "type": "object",
            "required": [
              "productId",
              "quantity",
              "_id"
            ],
            "properties": {
              "productId": {
                "description": "Product id",
                "type": "string",
                "example": "65e4decdd286b30065d54af9"
              },
              "quantity": {
                "description": "Product quantity",
                "type": "number",
                "example": 3
              },
              "_id": {
                "description": "Item id",
                "type": "string",
                "example": "65e4decdd286b30065d54af9"
              }
            }
          },
          "createdAt": {
            "description": "When the cart was created",
            "type": "string",
            "example": "2025-05-30T10:23:10.962+00:00"
          },
          "updatedAt": {
            "description": "When the cart was updated",
            "type": "string",
            "example": "2025-05-30T10:23:10.962+00:00"
          }
        }
      },
      "post": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "object",
            "required": [
              "productId",
              "quantity"
            ],
            "properties": {
              "productId": {
                "description": "Product id",
                "type": "string",
                "example": "65e4decdd286b30065d54af9"
              },
              "quantity": {
                "description": "Product quantity",
                "type": "number",
                "example": 3
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Invalid parameters"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong on the server"
                }
              }
            }
          }
        }
      }
    }
  }
}